Moved Matrix class
 - Moved Matrix class to its own module following @gfrances feedback
