+++ src/tarski/syntax/algebra/__init__.py
@@ -1,0 +1,1 @@
+from .matrix import Matrix
